Edit: change list type checking to float instead of int

+++ 0x00-python_variable_annotations/5-sum_list.py
@@ -6,7 +6,7 @@ argument and returns their sum as a float
 from typing import List
 
 
-def sum_list(input_list: List[int]) -> float:
+def sum_list(input_list: List[float]) -> float:
     """return sum of argument list elements"""
     total: float = 0.0
     for i in range(len(input_list)):
